Split player methods and inner classes out
Why are they all here and is this needed? No idea. Will it harm to move
them out? Looks like not.
