+++ src/main/emulator/debug/Memory.java
@@ -5,14 +5,12 @@ import com.samsung.util.AudioClip;
 import emulator.Emulator;
 import emulator.Settings;
 import emulator.graphics2D.IImage;
-import emulator.media.vlc.VLCPlayerImpl;
 import org.apache.tools.zip.ZipFile;
 
 import javax.microedition.lcdui.Image;
 import javax.microedition.m3g.*;
 import javax.microedition.media.Player;
 import javax.microedition.media.PlayerImpl;
-import javax.microedition.media.control.VolumeControlImpl;
 import java.io.*;
 import java.lang.reflect.Array;
 import java.lang.reflect.Field;
@@ -174,7 +172,7 @@ public final class Memory {
 			this.instances.add(o);
 			try {
 				if (o instanceof Image) {
-					this.images.add((Image)o);
+					this.images.add((Image) o);
 					if (Settings.recordReleasedImg && this.releasedImages.contains(o)) {
 						this.releasedImages.removeElement(o);
 					}
@@ -373,320 +371,6 @@ public final class Memory {
 		return n;
 	}
 
-	public static String getPlayerType(final Object o) {
-		if (o instanceof Sound) {
-			return ((Sound) o).getType();
-		}
-		if (o instanceof AudioClip) {
-			return "MMF";
-		}
-		return ((Player) o).getContentType();
-	}
-
-	public static String playerStateStr(final Object o) {
-		if (o instanceof Sound) {
-			switch (((Sound) o).getState()) {
-				case 0: {
-					return "SOUND_PLAYING";
-				}
-				case 1: {
-					return "SOUND_STOPPED";
-				}
-				case 3: {
-					return "SOUND_UNINITIALIZED";
-				}
-				default: {
-					return "INVALID STATE";
-				}
-			}
-		} else if (o instanceof AudioClip) {
-			switch (((AudioClip) o).getStatus()) {
-				case 1: {
-					return "SOUND_PLAY";
-				}
-				case 2: {
-					return "SOUND_PAUSE";
-				}
-				case 0: {
-					return "SOUND_STOP";
-				}
-				default: {
-					return "INVALID STATE";
-				}
-			}
-		} else {
-			switch (((Player) o).getState()) {
-				case 0: {
-					return "CLOSED";
-				}
-				case 300: {
-					return "PREFETCHED";
-				}
-				case 200: {
-					return "REALIZED";
-				}
-				case 400: {
-					return "STARTED";
-				}
-				case 100: {
-					return "UNREALIZED";
-				}
-				default: {
-					return "INVALID STATE";
-				}
-			}
-		}
-	}
-
-	public static int getPlayerLoopCount(final Object o) {
-		if (o instanceof Sound && ((Sound) o).m_player instanceof PlayerImpl) {
-			return ((PlayerImpl) ((Sound) o).m_player).loopCount;
-		}
-		if (o instanceof AudioClip) {
-			return ((AudioClip) o).loopCount;
-		}
-		if (o instanceof PlayerImpl) {
-			return ((PlayerImpl) o).loopCount;
-		}
-		return 0;
-	}
-
-	public static int getPlayerDurationMs(final Object o) {
-		if (o == null)
-			return -1;
-		if (o instanceof Sound) {
-			return getPlayerDurationMs(((Sound) o).m_player);
-		}
-		if (o instanceof AudioClip) {
-			return getPlayerDurationMs(((AudioClip) o).m_player);
-		}
-		if (o instanceof Player) {
-			long dur = ((Player) o).getDuration();
-			if (dur < 0) return -1;
-			return (int) (dur / 1000L);
-		}
-		return -1;
-	}
-
-	public static int getPlayerCurrentMs(final Object o) {
-		if (o == null)
-			return -1;
-		if (o instanceof Sound)
-			return getPlayerCurrentMs(((Sound) o).m_player);
-		if (o instanceof AudioClip)
-			return getPlayerCurrentMs(((AudioClip) o).m_player);
-		if (o instanceof Player) {
-			long l = ((Player) o).getMediaTime();
-			if (l < 0)
-				return -1;
-			return (int) (l / 1000L);
-		}
-		return -1;
-	}
-
-	public static int getPlayerDataLength(final Object o) {
-		if (o instanceof Sound) {
-			return ((Sound) o).dataLen;
-		}
-		if (o instanceof AudioClip) {
-			return ((AudioClip) o).dataLen;
-		}
-		if (o instanceof VLCPlayerImpl) {
-			return ((VLCPlayerImpl) o).dataLen;
-		}
-		if (!(o instanceof PlayerImpl)) {
-			return 0;
-		}
-		return ((PlayerImpl) o).dataLen;
-	}
-
-	public static int getPlayerVolume(final Object o) {
-		try {
-			if (o instanceof Sound) {
-				return ((Sound) o).getGain();
-			}
-			if (o instanceof AudioClip) {
-				return ((AudioClip) o).volume * 20;
-			}
-
-			if (o instanceof VLCPlayerImpl) {
-				return ((VolumeControlImpl) ((VLCPlayerImpl) o).getControl("VolumeControl")).getLevel();
-			}
-			if (o instanceof PlayerImpl) {
-				return ((VolumeControlImpl) ((PlayerImpl) o).getControl("VolumeControl")).getLevel();
-			}
-			return 0;
-		} catch (Exception ex) {
-			return 0;
-		}
-	}
-
-	public static void setPlayerVolume(final Object o, final int n) {
-		try {
-			if (o instanceof Sound) {
-				((Sound) o).setGain(n);
-			} else if (!(o instanceof AudioClip)) {
-
-				if (o instanceof VLCPlayerImpl) {
-					((VolumeControlImpl) ((VLCPlayerImpl) o).getControl("VolumeControl")).setLevel(n);
-					return;
-				}
-				if (!(o instanceof PlayerImpl)) {
-					return;
-				}
-				((VolumeControlImpl) ((PlayerImpl) o).getControl("VolumeControl")).setLevel(n);
-			}
-		} catch (Exception ignored) {
-		}
-	}
-
-	public static void modifyPlayer(final Object o, final PlayerActionType n) {
-		if (o instanceof Sound) {
-			final Sound sound = (Sound) o;
-			try {
-				switch (n) {
-					case resume: {
-						sound.resume();
-						break;
-					}
-					case pause: {
-						final long mediaTime = sound.m_player.getMediaTime();
-						sound.stop();
-						sound.m_player.setMediaTime(mediaTime);
-						break;
-					}
-					case stop: {
-						sound.stop();
-						break;
-					}
-					case export: {
-						try {
-							byte[] b = sound.getData();
-							String s = sound.getExportName();
-							if (b != null) {
-								exportAudio(b, sound.getExportName());
-								Emulator.getEmulator().getScreen().showMessage("Saved: " + s);
-							} else {
-								Emulator.getEmulator().getScreen().showMessage("Export failed: unsupported stream type");
-							}
-						} catch (Exception e) {
-							Emulator.getEmulator().getScreen().showMessage("Export failed: " + e);
-						}
-						break;
-					}
-				}
-			} catch (Exception ignored) {
-			}
-			return;
-		}
-		if (o instanceof AudioClip) {
-			final AudioClip audioClip = (AudioClip) o;
-			switch (n) {
-				case resume: {
-					audioClip.play(audioClip.loopCount, audioClip.volume);
-					break;
-				}
-				case pause: {
-					audioClip.pause();
-					break;
-				}
-				case stop: {
-					audioClip.stop();
-					break;
-				}
-				case export: {
-					try {
-						byte[] b = audioClip.getData();
-						String s = audioClip.getExportName();
-						if (b != null) {
-							exportAudio(b, s);
-							Emulator.getEmulator().getScreen().showMessage("Saved: " + s);
-						} else {
-							Emulator.getEmulator().getScreen().showMessage("Export failed: unsupported stream type");
-						}
-					} catch (Exception e) {
-						Emulator.getEmulator().getScreen().showMessage("Export failed: " + e);
-					}
-					break;
-				}
-			}
-			return;
-		}
-		if (o instanceof VLCPlayerImpl) {
-			final VLCPlayerImpl v = (VLCPlayerImpl) o;
-			try {
-				switch (n) {
-					case resume: {
-						v.start();
-						break;
-					}
-					case pause: {
-						final long mediaTime2 = v.getMediaTime();
-						v.stop();
-						v.setMediaTime(mediaTime2);
-					}
-					case stop: {
-						v.stop();
-						break;
-					}
-					case export: {
-						Emulator.getEmulator().getScreen().showMessage("Export not supported!");
-						break;
-					}
-				}
-			} catch (Exception ex2) {
-				ex2.printStackTrace();
-			}
-			return;
-		}
-		if (!(o instanceof PlayerImpl)) {
-			return;
-		}
-		final PlayerImpl playerImpl = (PlayerImpl) o;
-		try {
-			switch (n) {
-				case resume: {
-					playerImpl.start();
-					break;
-				}
-				case pause: {
-					final long mediaTime2 = playerImpl.getMediaTime();
-					playerImpl.stop();
-					playerImpl.setMediaTime(mediaTime2);
-					break;
-				}
-				case stop: {
-					playerImpl.stop();
-					break;
-				}
-				case export: {
-					try {
-						byte[] b = playerImpl.getData();
-						String s = "audio" + playerImpl.getExportName();
-						if (b != null) {
-							exportAudio(b, s);
-							Emulator.getEmulator().getScreen().showMessage("Saved: " + s);
-						} else {
-							Emulator.getEmulator().getScreen().showMessage("Export failed: unsupported stream type");
-						}
-					} catch (Exception e) {
-						Emulator.getEmulator().getScreen().showMessage("Export failed: " + e);
-					}
-				}
-			}
-		} catch (Exception ignored) {
-		}
-	}
-
-	public static void exportAudio(byte[] b, String name) throws IOException {
-		File f = new File(Emulator.getUserPath() + "/" + name);
-		if (f.exists()) return;
-		f.createNewFile();
-		DataOutputStream o = new DataOutputStream(new FileOutputStream(f));
-		o.write(b);
-		o.close();
-	}
-
 	public final int instancesCount(final String o) {
 		try {
 			return this.classesTable.get(o).instancesCount;
@@ -808,45 +492,5 @@ public final class Memory {
 			throw new NoClassDefFoundError(ex2.getMessage());
 		}
 		return forName;
-	}
-
-	public static final class ObjInstance {
-		public final String reference;
-		public final Object value;
-		public final int size;
-
-		ObjInstance(final Memory a, final String ref, final Object o) {
-			super();
-			this.reference = ref;
-			this.value = o;
-			this.size = a.size(o.getClass(), o);
-		}
-	}
-
-	public static final class ClassInfo implements Comparable<ClassInfo> {
-		final String name;
-		int instancesCount;
-		final int staticsSize;
-		public final Vector<ObjInstance> objs;
-
-		public int size() {
-			int total = this.staticsSize;
-			for (int i = this.objs.size() - 1; i >= 0; --i) {
-				total += this.objs.get(i).size;
-			}
-			return total;
-		}
-
-		public int compareTo(final ClassInfo o) {
-			return this.name.compareTo(o.name);
-		}
-
-		ClassInfo(final Memory m, final Class cls) {
-			super();
-			this.objs = new Vector();
-			this.instancesCount = 0;
-			this.staticsSize = m.size(cls, null);
-			this.name = cls.getName();
-		}
 	}
 }
